refactor: extract focused values into ContentViewFocusedValues modifier
Moves ~130 lines of .focusedValue() calls and the folder-watch
options sheet from ContentView into a dedicated ViewModifier.

Part of #284
